% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_cpm.R
\name{countToCPM}
\alias{countToCPM}
\title{Generate TPM value using Gene ReadCount data}
\usage{
countToCPM(expmat)
}
\arguments{
\item{expmat}{A dataframe generated from 'geneLength()', with 'est_len' column}
}
\value{
Dataframe
}
\description{
Generate TPM value using Gene ReadCount data
}
