% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_fpkm.R
\name{countToFpkm}
\alias{countToFpkm}
\title{Generate FPKM value using Gene ReadCount data}
\usage{
countToFpkm(expmat, effLen_col)
}
\arguments{
\item{expmat}{A dataframe generated from 'geneLength()', with 'est_len' column}

\item{effLen_col}{Specify the column representing gene lengths in the input data frame}
}
\value{
Dataframe
}
\description{
Generate FPKM value using Gene ReadCount data
}
