% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deseq2_analysis.R
\name{desCompare}
\alias{desCompare}
\title{DEGs betweent two groups}
\usage{
desCompare(
  readcount = NULL,
  group1.name = NULL,
  group1.number = NULL,
  group2.name = NULL,
  group2.number = NULL,
  log2FC.threshold = 1,
  pvalue.threshold = 0.05,
  padj.threshold = 0.05,
  top.n = 10,
  dataset = NULL,
  type = "ENSEMBL"
)
}
\arguments{
\item{readcount}{A data frame containing ReadCount, with rows representing gene names and columns representing sample names.}

\item{group1.name}{Custom group names, such as "OV".}

\item{group1.number}{The number of samples in Group 1.}

\item{group2.name}{Custom group names, such as "Control".}

\item{group2.number}{The number of samples in Group 2.}

\item{log2FC.threshold}{Set the threshold for the absolute value of log2 Fold Change (|log2FC|) for differentially expressed genes, with a default value of 1.}

\item{pvalue.threshold}{The p-value for differentially expressed genes, with a default of 0.05.}

\item{padj.threshold}{The adjusted p-value for differentially expressed genes, with a default of 0.05.}

\item{top.n}{The number of gene names ranked at the top (with the smallest p-value and the largest log2FC).}

\item{dataset}{Results from the refer function.}

\item{type}{The type of gene names in the ReadCount file: ENSEMBL or SYMBOL.}
}
\value{
List
}
\description{
Conducting differential analysis using DESeq2 with ReadCount.
}
\examples{
res_list <- desCompare(readcount=data, group1.name="A", group1.number=3, group2.name="B", group2.number=3,dataset=dataset, type="ENSEMBL")
}
