% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRNA.R
\docType{package}
\name{easyRNA}
\alias{easyRNA}
\alias{easyRNA-package}
\title{easyRNA: Custom Functions for mRNA-seq Experiments}
\description{
A set of costum function to the analysis of mRNA-seq data sets.
}
