% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findname.R
\name{findname}
\alias{findname}
\title{Find Another Type of Name}
\usage{
findname(query, dataset, type)
}
\arguments{
\item{query}{An object which want to query}

\item{dataset}{A dataset constructed from GTF: generated by 'refer' function}

\item{type}{Type of input gene name ENSEMBL or SYMBOL}
}
\value{
A global object which transformed name.
}
\description{
Find another name of same gene.
}
\examples{
result <- findname("ENSGALG00000017145",dataset, "ENSEMBL")


}
