% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm_to_tpm.R
\name{fpkmToTpm}
\alias{fpkmToTpm}
\title{Generate TPM value using FPKM}
\usage{
fpkmToTpm(expmat)
}
\arguments{
\item{expmat}{}
}
\value{
Dataframe
}
\description{
Generate TPM value using FPKM
}
