% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_length.R
\name{geneLength}
\alias{geneLength}
\title{A GeneLength Used for Counting}
\usage{
geneLength(df, gtf)
}
\arguments{
\item{df}{Here is the ReadCount data for each sample, with gene names as row names and sample names as column names.}

\item{gtf}{Input gtf.file for your species}
}
\value{
Return a data frame containing the corresponding gene lengths and the ReadCount for each sample.
}
\description{
Before transform the count into other, a gene length should be made.
}
\examples{
length <- geneLength(gtf)
}
