% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{getDE}
\alias{getDE}
\title{Get Differentially Expressed Genes by Custum Significance Threshold}
\usage{
getDE(file_name, fold_change, p_value, path = "./data/raw/")
}
\arguments{
\item{file_name}{file name in the data folder}

\item{fold_change}{fold change threshold}

\item{p_value}{p-value threshold}

\item{path}{path to the DESeq files}
}
\value{
vector of differentially expressed genes
}
\description{
\code{getDE} read the DESeq output file and creat or rewrite the DE column based on the manualy set fold change and p-value thesholds.
}
\examples{

}
