% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_score.R
\name{getRES}
\alias{getRES}
\title{Calculate Running Enrichment Score to Any Gene Set}
\usage{
getRES(target_set, ranked_universe)
}
\arguments{
\item{target_set}{vactor of ensembl gene ids}

\item{ranked_universe}{DEseq2 output file with gene rank column}
}
\value{
Running Enrichment Score (RES) plot
}
\description{
\code{getRES}
}
\examples{
getRES(c(1,2), c(1,2,3,4,5,6,7))
getRES(c('a','b'), c('a','b','c','d','e','f','g'))
}
