% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_hypergeometric_test.R
\name{hgTest}
\alias{hgTest}
\title{Run Hypergeometr Test On a Set of Target Genes}
\usage{
hgTest(target_set, gene_universe, outputN)
}
\arguments{
\item{target_set}{vector of target ensembl ids}

\item{gene_universe}{vector of universe ensembl ids}

\item{outputN}{file name of the ouptut}
}
\value{

}
\description{
Run Hypergeometr Test On a Set of Target Genes
}
\examples{

}
