% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA analysis of bulk RNA-seq}
\usage{
pca_plot(
  data,
  groups,
  group_positions,
  point.size = 4,
  label.size = 4,
  colors,
  ellipse = FALSE
)
}
\arguments{
\item{data}{A data frame of the TPM or FPKM values of genes, with row names as gene names and column names as sample names.}

\item{groups}{Group names, in the form of a vector.}

\item{group_positions}{A list for the distribution of samples of each group in the data frame.}

\item{point.size}{point size. Default:4}

\item{label.size}{label size. Default:4}

\item{colors}{point colors}

\item{ellipse}{Whether to draw a confidence ellipse. Default:FALSE}
}
\description{
PCA analysis of bulk RNA-seq
}
\examples{
pca_plot(data = data_tpm, groups = groups, group_positions = group_positions, colors = colors, ellipse = TRUE)
}
