% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corelation_heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Plot correlation heatmap}
\usage{
plot_correlation_heatmap(
  data,
  method = "pearson",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  sample_annotation = NULL,
  annotation_colors = NULL,
  heatmap_colors = colorRampPalette(c("blue", "white", "red"))(100),
  show_numbers = TRUE,
  main_title = "Sample Correlation Heatmap"
)
}
\arguments{
\item{data}{A data frame of the TPM or FPKM values of genes, with row names as gene names and column names as sample names.}

\item{method}{Correlation calculation method: "pearson", "spearman", "kendall"}

\item{cluster_rows}{Whether to cluster rows}

\item{cluster_cols}{Whether to cluster the column}

\item{sample_annotation}{Sample annotation data frame. The row names are sample names, and the column names are different traits or groups.}

\item{annotation_colors}{Sample Annotation Color. Specify a list of colors for members in different groups, used for legend colors}

\item{heatmap_colors}{Heatmap color scheme}

\item{show_numbers}{Whether to display the correlation coefficient value}

\item{main_title}{Heatmap title}
}
\value{
ggplot2
}
\description{
Plot correlation heatmap
}
\examples{
plot_correlation_heatmap(data = data_tpm, method = "spearman", sample_annotation = sample_annotation, annotation_colors = list(Group=c("A"="red3","B"="blue3"),Gender=c("Male"="purple", "Female"="yellow2")))
}
