% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_violin.R
\name{plot_gene_violin}
\alias{plot_gene_violin}
\title{Plot gene viollin}
\usage{
plot_gene_violin(
  data,
  groups,
  group_positions,
  gene_of_interest,
  method = "t.test",
  colors = NULL
)
}
\arguments{
\item{data}{A data frame of the TPM or FPKM values of genes, with row names as gene names and column names as sample names.}

\item{groups}{Group names, in the form of a vector.}

\item{group_positions}{A list for the distribution of samples of each group in the data frame.}

\item{gene_of_interest}{Gene of interest.}

\item{method}{Methods for Differential Analysis Between Groups.}

\item{colors}{The color of each group.}
}
\description{
Plot gene viollin
}
\examples{
plot_gene_violin(data = data_tpm, groups = c("A","B"), group_positions = list(1:3,4:6), gene_of_interest = c("ENSGALG00000012172"), method = "t.test")
}
