% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot volcano using output from desCompare}
\usage{
plot_volcano(
  deg.data,
  title = "Volcano Plot",
  xlab = "log2FoldChange",
  ylab = "-log10(P-value)",
  x_cutoff = c(-1, 1),
  y_cutoff = 1.3,
  label_column = "label",
  group_column = "Group",
  log_scale = "log10p",
  point_size = 2,
  label_size = 3,
  colors = c("purple", "#BBBBBB", "orange")
)
}
\arguments{
\item{deg.data}{desCompare's DEG_Dataframe}

\item{title}{volcano's title}

\item{xlab}{xlab's title. Default:log2FoldChange}

\item{ylab}{ylab's title. Default:-log10(P-value)}

\item{x_cutoff}{The threshold line on the X-axis.}

\item{y_cutoff}{The threshold line on the Y-axis.}

\item{label_column}{"label" column. Do not change.}

\item{group_column}{"Group" column. Do not change.}

\item{log_scale}{Should points be plotted using log10p or log10q for Y-axis values? Default: log10p.}

\item{point_size}{Default:2}

\item{label_size}{Default:3}

\item{colors}{point colors. Default:colors = c("purple", "#BBBBBB", "orange")}
}
\value{
volcano plot
}
\description{
plot volcano using output from desCompare
}
\examples{
plot_volcano(deg.data = res[["DEG_Dataframe"]], title = "HC:AA_vs_XH", log_scale="log10p")
}
