% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rMATS_results.R
\name{process_rMATS_results}
\alias{process_rMATS_results}
\title{process_rMATS_results}
\usage{
process_rMATS_results(
  event_files,
  pvalue_cutoff = 0.05,
  diff_cutoff = 0.1,
  orgdb = org.Hs.eg.db,
  go_analysis = TRUE,
  kegg_analysis = TRUE,
  save_results = TRUE,
  output_prefix = "rMATS_results"
)
}
\arguments{
\item{event_files}{A list showing the Alternative Splicing Event files}

\item{pvalue_cutoff}{p-value threshold for differential splicing events}

\item{diff_cutoff}{PSI threshold for differential splicing events}

\item{orgdb}{GO and KEGG enrichment analysis databases. Default: org.Hs.eg.db}

\item{go_analysis}{Whether to conduct GO analysis}

\item{kegg_analysis}{Whether to conduct  KEGG analysis}

\item{save_results}{Whether to save the result in csv file}

\item{output_prefix}{Output file prefix}
}
\description{
process_rMATS_results
}
\examples{
event_files <- list(SE = "SE.MATS.JC.txt",RI = "RI.MATS.JC.txt",MXE = "MXE.MATS.JC.txt",A5SS = "A5SS.MATS.JC.txt",A3SS = "A3SS.MATS.JC.txt");
process_rMATS_results(event_files)
}
