% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refer.R
\name{refer}
\alias{refer}
\title{Make a gtf database}
\usage{
refer(gtf)
}
\arguments{
\item{gtf}{A gtf for chicken from ENSEMBL}
}
\value{
a global object df
}
\description{
This function is used to construct a database from gtf file.
}
\examples{
dataset <- refer("Gallus_gallus.GRCg6a.10.gtf")

}
