% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GO_analysis.R
\name{run_GO_analysis}
\alias{run_GO_analysis}
\title{GO analysis and visulization}
\usage{
run_GO_analysis(
  des_res,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  topn = 10,
  up_color = "purple",
  down_color = "orange",
  font_size = 3,
  title_size = 14
)
}
\arguments{
\item{des_res}{The results of differentially expressed genes from desCompare function.}

\item{pvalueCutoff}{Threshold for GO analysis. Default:0.05}

\item{qvalueCutoff}{Threshold for GO analysis. Default:0.05}

\item{topn}{How many GO items should be displayed?}

\item{up_color}{GO term color of up-regulated genes}

\item{down_color}{GO term color of down-regulated genes}

\item{font_size}{Font size of GO entries. Default:12}

\item{title_size}{Title size}
}
\value{
ggplot2
}
\description{
GO analysis and visulization
}
\examples{
run_GO_analysis(des_res = des_res, topn = 15, up_color = "red4", down_color = "green4", font_size = 12)
}
