% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KEGG_analysis.R
\name{run_KEGG_analysis}
\alias{run_KEGG_analysis}
\title{KEGG analysis}
\usage{
run_KEGG_analysis(
  des_res,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  up_color = "purple",
  down_color = "orange",
  font_size = 10,
  point_size = 5
)
}
\arguments{
\item{des_res}{The results of differentially expressed genes from desCompare function.}

\item{pvalueCutoff}{Threshold for KEGG analysis. Default:0.05}

\item{qvalueCutoff}{Threshold for KEGG analysis. Default:0.05}

\item{up_color}{KEGG term color of up-regulated genes}

\item{down_color}{KEGG term color of up-regulated genes}

\item{font_size}{Font size of KEGG items}

\item{point_size}{Point Size}
}
\value{
ggplot2
}
\description{
KEGG analysis
}
\examples{
run_KEGG_analysis(des_res, pvalueCutoff = 0.2, qvalueCutoff = 0.2, up_color = "blue", down_color = "red", font_size =2, point_size = 8)
}
