% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_cpm.R
\name{countToCPM}
\alias{countToCPM}
\title{Change Count into CPM}
\usage{
countToCPM(counts)
}
\arguments{
\item{counts}{Gene Count}
}
\value{
CPM value
}
\description{
Change Count into CPM
}
\examples{
df \%>\%mutate(CPM=countToCPM(.$count))
}
