% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_fpkm.R
\name{countToFpkm}
\alias{countToFpkm}
\title{Change Count into FPKM}
\usage{
countToFpkm(counts, effLen)
}
\arguments{
\item{counts}{Gene Count}

\item{effLen}{Gene Length}
}
\value{
FPKM value
}
\description{
Change Count into FPKM
}
\examples{
df \%>\%mutate(FPKM=countToFpkm(.$count,.$est_len))
}
