% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_tpm.R
\name{countToTpm}
\alias{countToTpm}
\title{Change Gene Count into TPM}
\usage{
countToTpm(counts, effLen)
}
\arguments{
\item{counts}{Gene Count}

\item{effLen}{Gene Length}
}
\value{
TPM value
}
\description{
Change Gene Count into TPM
}
\examples{
df \%>\% mutate(TPM=countToTpm(.$count, .$est_len))
}
