% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_tpm.R
\name{countToTpm}
\alias{countToTpm}
\title{Generate TPM value using Gene ReadCount data}
\usage{
countToTpm(expmat, effLen_col)
}
\arguments{
\item{expmat}{A dataframe generated from 'geneLength()', with 'est_len' column}

\item{effLen_col}{Specify the column representing gene lengths in the input data frame}
}
\value{
Dataframe
}
\description{
Generate TPM value using Gene ReadCount data
}
