% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm_to_tpm.R
\name{fpkmToTpm}
\alias{fpkmToTpm}
\title{Change fpkm into tpm}
\usage{
fpkmToTpm(fpkm)
}
\arguments{
\item{fpkm}{input the fpkm value}
}
\value{
TPM value
}
\description{
Change fpkm into tpm
}
\examples{
df \%>\% mutate(TPM=fpkmToTpm(.$fpkm))
}
