% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_length.R
\name{geneLength}
\alias{geneLength}
\title{A GeneLength Used for Counting}
\usage{
geneLength(df, gtf)
}
\arguments{
\item{df}{A combined dataframe with two column:gene_id and count. One of the column must be 'gene_id'}

\item{gtf}{Input gtf for your species}
}
\value{
A dataframe
}
\description{
Before transform the count into other, a gene length should be made.
}
\examples{
length <- geneLength(gtf)
}
