% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{getDE}
\alias{getDE}
\title{Get Differentially Expressed Genes by Custum Significance Threshold}
\usage{
getDE(fileN, fct, pvt)
}
\arguments{
\item{fileN}{file name in the data folder}

\item{fct}{fold change threshold}

\item{pvt}{p-value threshold}
}
\value{

}
\description{
\code{getDE} read the DESeq output file and creat or rewrite the DE column based on the manualy set fold change and p-value thesholds.
}
\examples{

}
