% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_score.R
\name{getRES}
\alias{getRES}
\title{Plot Running Enrichment Score to Any Gene Set}
\usage{
getRES(target_set, universe)
}
\arguments{
\item{target_set}{vactor of ensembl gene ids}

\item{universe}{DEseq2 output file with gene rank column}
}
\value{
Running Enrichment Score (RES) plot
}
\description{
\code{getRES}
}
\examples{

}
